client/src/main/java/io/steem/client/model/operation/WitnessUpdateOperation.java
package io.steem.client.model.operation;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.databind.PropertyNamingStrategy;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import io.steem.client.model.SteemAsset;
import io.steem.client.model.SteemChainProperties;
import io.steem.client.model.SteemOperation;
import io.steem.client.model.util.ObjectMapUtils;
import lombok.Builder;

import java.util.Map;

public class WitnessUpdateOperation extends SteemOperation {

  @Builder
  @lombok.Value
  @JsonInclude(JsonInclude.Include.NON_NULL)
  @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
  public static class Value {
    private final String owner;
    private final String url;
    private final String blockSigningKey;
    private final SteemChainProperties props;
    private final SteemAsset fee;
  }

  private final Value value;

  private WitnessUpdateOperation(Value value) {
    super("witness_update");
    this.value = value;
  }

  public static WitnessUpdateOperation of(Value value) {
    return new WitnessUpdateOperation(value);
  }

  @Override
  protected Map<String, Object> getValueMap() {
    return ObjectMapUtils.toObjectMap(value);
  }
}
